backend/requirements.txt
fastapi
uvicorn
bcrypt
passlib
python-jose[cryptography]
pyjwt
sqlalchemy
databases
requests
azure-cosmos
azure-identity
azure-mgmt-resource
azure-mgmt-compute
azure-mgmt-storage
azure-mgmt-network
azure-mgmt-monitor
azure-mgmt-loganalytics
azure-mgmt-recoveryservices
azure-mgmt-subscription
azure-mgmt-automation
azure-mgmt-logic
jinja2
itsdangerous
python-dotenv